+++ lib/mayaUsd/nodes/proxyShapeUpdateManager.h
@@ -1,0 +1,47 @@
+//
+// Copyright 2023 Autodesk
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+#ifndef PXRUSDMAYA_PROXY_SHAPE_UPDATE_MANAGER_H
+#define PXRUSDMAYA_PROXY_SHAPE_UPDATE_MANAGER_H
+
+#include <mayaUsd/base/api.h>
+
+#include <pxr/pxr.h>
+#include <pxr/usd/usd/notice.h>
+
+#include <maya/MStatus.h>
+
+PXR_NAMESPACE_OPEN_SCOPE
+
+class MayaUsdProxyShapeUpdateManager
+{
+public:
+    MAYAUSD_CORE_PUBLIC
+    bool CanIgnoreObjectsChanged(const UsdNotice::ObjectsChanged& notice);
+
+    MAYAUSD_CORE_PUBLIC
+    MInt64 GetUpdateCount();
+
+    MAYAUSD_CORE_PUBLIC
+    MInt64 GetResyncCount();
+
+private:
+    MInt64 _UsdStageUpdateCounter { 1 };
+    MInt64 _UsdStageResyncCounter { 1 };
+};
+
+PXR_NAMESPACE_CLOSE_SCOPE
+
+#endif
