+++ test/lib/ufe/testUINodeGraphNode.py
@@ -60,8 +60,8 @@ class UINodeGraphNodeTestCase(unittest.TestCase):
         ball3Path = ufe.PathString.path('|transform1|proxyShape1,/Ball_set/Props/Ball_3')
         ball3SceneItem = ufe.Hierarchy.createItem(ball3Path)
 
-        initialUpdateCount = cmds.getAttr('|transform1|proxyShape1.updateid')
-        initialResyncCount = cmds.getAttr('|transform1|proxyShape1.resyncid')
+        initialUpdateCount = cmds.getAttr('|transform1|proxyShape1.updateId')
+        initialResyncCount = cmds.getAttr('|transform1|proxyShape1.resyncId')
                                           
         uiNodeGraphNode = ufe.UINodeGraphNode.uiNodeGraphNode(ball3SceneItem)
 
@@ -89,8 +89,8 @@ class UINodeGraphNodeTestCase(unittest.TestCase):
         self.assertEqual(pos4.y(), pos5.y())
 
         # None of these changes should force a render refresh:
-        self.assertEqual(initialUpdateCount, cmds.getAttr('|transform1|proxyShape1.updateid'))
-        self.assertEqual(initialResyncCount, cmds.getAttr('|transform1|proxyShape1.resyncid'))
+        self.assertEqual(initialUpdateCount, cmds.getAttr('|transform1|proxyShape1.updateId'))
+        self.assertEqual(initialResyncCount, cmds.getAttr('|transform1|proxyShape1.resyncId'))
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
